+++ CHANGELOG.md
@@ -1,4 +1,22 @@
 ## CHANGELOG
+### 0.5.0
+Major Update again, sorry for the big changes. This has been a rough road, but I feel it has finally smoothed out and I'm very happy with where everything is at.
+
+**Name Change**
+`react-motion-slider` -> `react-view-pager`
+
+**Breaking Changes**
+*Props changed*
+`currentIndex` && `currentKey` -> `currentView`
+`slidesToShow` -> `viewsToShow`
+`slidesToMove` -> `viewsToMove`
+`vertical` -> `axis`
+`autoHeight` -> `autoSize`
+
+*Props added*
+`contain`: don't allow empty space in frame
+`lazyLoad`: load components on the fly as they enter the frame
+
 ### 0.4.2
 Use constructor in `Slider` to fix undefined props in IE10